Convert modal handlers to typescript
